+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.12.1
 * Spawning feature (renamed from Base) can now be disabled
+* Creeper cena now also breaks more blocks if SR is present
 
 ## 1.12.0
 * Added Biting Zombies feature